attached_assets/requirements_1756363462142.txt
flask==3.0.3
dnspython==2.6.1
selenium==4.21.0
requests==2.32.3
gunicorn==22.0.0
flask-session==0.5.0
python-decouple==3.8
pillow==10.4.0
flask-wtf==1.2.1
pyjwt==2.8.0
bcrypt==4.1.3
flask-limiter==3.5.1
redis==5.0.1
sentry-sdk==2.7.1
pytest==8.2.2
flask-sqlalchemy==3.1.1